Fix UR_DEVICE_INFO_SUB_GROUP_SIZES_INTEL size confusion
The spec says it's an array of uint32_t, but CUDA, HIP, OpenCL, and
LevelZero adapters were returning arrays of size_t. As a result, we had
extra dummy zeroes in the output. Native CPU adapter was correct.

+++ source/adapters/cuda/device.cpp
@@ -270,7 +270,7 @@ UR_APIEXPORT ur_result_t UR_APICALL urDeviceGetInfo(ur_device_handle_t hDevice,
     int WarpSize = 0;
     UR_CHECK_ERROR(cuDeviceGetAttribute(
         &WarpSize, CU_DEVICE_ATTRIBUTE_WARP_SIZE, hDevice->get()));
-    size_t Sizes[1] = {static_cast<size_t>(WarpSize)};
+    uint32_t Sizes[1] = {static_cast<uint32_t>(WarpSize)};
     return ReturnValue(Sizes, 1);
   }
   case UR_DEVICE_INFO_MAX_CLOCK_FREQUENCY: {
